[REF] simplify references to civicrm_acl
We can see looking at ACL.xml that civicm_acl can't be localised so we gain nothing
by using a varible.

Also note that we want to switch to CRM_Core_Execute::query in these functions so
that table translation is handled by that function - this is old code done the old way.

However, the thing I'm trying to unravel is if & where the values for the fields
entity_table & object_table are localised. We have been switching AWAY from
saving localised table names & TBH I think we maybe don't & it's a lot of cruft
for some ideas that never eventuated but I'm still digging