Remove calls & deprecate CRM_Core_BAO_PrevNextCache::setItem
We have this function which

1) is mildly misleading - it appears to be generic to the prevnext table but is actually
only relevant to deduping as searches no longer use it
2) is about 50% deprecated and
3) the remainder is a single insert
4) is called from 2 places which use it a little differently

I think it's not really adding much value - I was going to remove the deprecated code
but I think in fact the goal should be to remove the whole function.

In general I think code on CRM_Core_BAO_PrevNextCache that is really dedupe-only code
should be on the dedupe classes. The history is that prevnext was created for searches
and kindof twisted to support dedupe as well but now search doesn't use much of what
is in the BAO class (if any)