Skip expensive smarty Processing
On digging into speed & disk use issues I find that significant disk space is being
used on smartyProcessing of greetings. The same also seems to be true of display names. This is
a first shot across the bows but I'm  also investigating
1) also doing an early return before replaceGreetingTemplate
2) changing  the   smarty ->fetch call to pass 'eval' rather than  string.
Despite the gut-reaction the docs identify this  as correct
https://www.smarty.net/docs/en/resources.string.tpl and eval is not eval

From my local profiling this  code also kicks in  on delete & it's likely we will actually
see a decrease in overall test time on this