Do not create smarty cached templates for processed greetings
Currrently whenever a greeting is parsed through smarty a smarty template file is stored to disk. This has
impacts on disk use, cache clearing time and performance. Where we know that the template will be of low value
we are better not to cache it. Smarty provides eval (not that eval) for this

The technical details are in the code comment block