dev/core#1630 Partial fix on renew membership missing
I'm pretty comfortable that this regressed at least  a year ago - possibly much longer - so it's not
a recent regression. However, the issue is that when a price set is configured it is still 'listening to'
the  presence or otherwise of auto_renew in civicrm_membership_block.membership_types. This makes it
ignore that if not present (which is the case when creating a new price set from scratch).
When converting a price set the membership_types field is still populated and still kicks in
- I think we should really empty that out when a block is converted but I've left it out of scope as this
seems like an edge case bug & it was only the possibity it was a regression that caused me to dig into it