Add more support for defining a pseudoconstant for a setting
This extends the ways in which a pseudoconstant can be defined in a setting to better reflect the  ways that
work for the DAO objects.

In this one field is converted - default_invoice_page under CiviContribute settings. (Described as
Default invoice payment page)

The expected result is that the options in that page load the same as before.

This change will reduce the temptation to call silly core functions from extensions....