[REF] Make apiv3-specific helper generally available.
This function should only be called from apiv3 but it's prevalence suggests it's more
generally useful. I have only done a couple of sample switches to use it. I think the
apiv4 one could in theory result in an error as is since there is no require_once for the
utils.php file