Fix MessageTemplates to be retrievable from workflow-id->option_value.name
This adds support for translating workflow_id to name or label in it's current schema & makes it easier for
us to modify the schema to use our normal structure.

Note I implemented the handling in buildOptions as putting a psuedoconstant callback
in doesn't allow different contexts. The hope is that with the code no longer
relying on the different option_groups being used we can do an upgrade script to
fix the structure to be a single option group and to refer to the value
field like others do