Call completeOrder directly from contribute status update form
We should really call completetransaction but that has been left for later unravelling with the focus here being
to clarify the purpose of the transaction. We don't need to pass a transaction around unless it has DB changes not
yet committed - which is not eally the case here. There is no reason to invoke BaseIPN but more unravelling required...