Fix pricefield pseudoconstant.
The price field turns out not to be returning pseudoconstant fields for 2 reasons
1) the label field is defined but the name field is not
2) any price fields with no domain are filtered out. In my single domain db this is
all of them.

I think we can safely understand NULL domain_id as 'not restricted by domain'. For some tables
(membershipType) it is a required field but for PriceField it is not.

I would be inclined to say it should be required less often. However, where it is required thenn
the OR IS NULL should never be true