Add more search options
This adds more fields to the filters on search for individual
- organization_name (Employer Name)
- first_name
- nick_name
- last_name

Note
1) organization_name is populated for individuals with employers - it's handy to be able
to filter by as a quick way to narrow down. I wasn't too worried about declaring it only
for individuals as sort_name covers this for organizations
2) Last name these feels a bit unnecessary as it co-incides with sort_name
but it also feels odd not to add it
3) Ideally for these text fields they would be always or optionally 'like' searches
(respecting the site +wildcard setting). I took a a quick look but decided to leave it
out of scope at this stage.
4) I thought 'email' might have been useful too but given 3 & the fact that the main use I saw
was as a proxy for 1 (employer) and that email is the preferred search for email anyway (
which was the entity I was working on) I excluded it too in the end.

The use case for this is narrowing down in a large database. Take finding Johnathan Smith.
Being able to add 'Johnathan' as a filter helps get past all the Johns quicker.

Lastly - I have a sneaky feeling that users in our org will abuse the nick name filter heavily &
add nicknames to all contacts they frequently email....