Remove validation bypass
I was trying to understand why executeQuery allows validation to be bypassed. That seems
intuitively like something that should be handled by the calling function. This is one of the few
places that calls it - but I can't see a reason for it to by pass validation....

My best guess is that a parameter could legitimately be NULL - and the group properly is the
only one that seems to be a candidate

I also note the case for passing i18nRewrite seems dubious as the table has no localisation