APIv4 - Support "relative date ranges" with the "IS" operator
Before
------

There's no good way to use the "relative date ranges" (e.g. "previous.month", "previous.year", "this.fiscal_year").

After
-----

This adds a new APIv4 query operator, "IS". This is loosely similar to the SQL "IS" operator
in that it can take a symbolic value (as in `myfield is null`). This flavor of "IS" accepts
"null" as well as the relative date ranges, e.g.

```php
civicrm_api4("Activity", "get", [
  "where" => [["activity_date_time", "IS", "previous.year"]],
  "select" => ["id","activity_date_time"],
  "limit" => 10,
]);
```