Move Legacy FirstData processor to an core-extension
This is one of a handful of processors that is in core but we believe is no longer in use.

The intent is the upgrade script should
1) Look for any instances of PaymentProcessors linked to the FirstData class. If none exist
delete the FirstData row from the civicrm_payment_processor_type table
2) Look for any financial_trxn records linked to any existing FirstData payment processor instance
(recurrings are not possible here). If none exist delete from civicrm_payment_processor & civicrm_payment_processor_type
3) If there IS a processor row then install the extension. The install makes no DB changes as this
extension is only intended for those who already depend on it existing. As such it is also hidden.

Planning for complete removal from core is punted for now. However, we should start to see stats on whether it is enabled
for anyone over time