Update guzzle, psr, psr-log & civicrm/civicrm-cxn-rpc
This brings guzzle up to the version of Drupal 8.

There is a minor compatibility issue between this version of psr7 & the previous version in the
error handling. This allows extensions that ship with vendor files to choose a sensible versionn.

Note extensions that ship with vendor packages can't exclude guzzle as it is a dependency for other packages