Fix payment instrument bug by using correct payment instrument
Per https://github.com/civicrm/civicrm-core/pull/17589 it is necessary that we do NOT pass the wrong payment_instrument_id
to the payment processor. BUT we have to provide accurate defaults for manual payments. This is a narrower fix than

- back office membership form
- back office participant form
- back office contribution form

and in all cases the Manual processor still loaded correctly. I don't feel I can say this is the last fix in this
area but it stands alone as a sensible fix to do and also one that should address the immediate issue.

Note there is some weirdness in a second function in EventFees - that function should GO IMHO - but I have not yet reached it in
UI testing to confirm if other changes need to be made.