dev/core#1603 Add tpl level formattinng for tax_amount
This is a partial fix for https://lab.civicrm.org/dev/core/-/issues/1603 because
we should be doing our formatting at the presentation layer rather than the BAO layer.
I suspect the wrong values might still be saved in the database but this gives us a bit
more space to fix that without breakin the display.

It also includes the brickMoney library per https://github.com/civicrm/civicrm-core/pull/17608
as part of the fix. Note that the rounding I chose was RoundingMode::CEILING
- this is not set in stone but the sum of the individual amounts come closer
to the total than with rounding down.

Note there were 2 params passed to the function that I removed - I think they are unnused but
I am digging further. Also note that this uses the default locale but we might prefer to
pass the user's locale when we know it as the locale can alter formatting e.g for Euro