[REF] - Add helper function for the repetitive task of fetching multilingual
Adds the function CRM_Core_I18n::getMultilingual() and replaces a bunch of
places where the work was being done manually.

This also fixes a suspected bug where multiple domains may not have been compatible
with multiple languages, as the current domain id was previously ignored,
but is respected in the helper function.