SetupController - Need better bootstrap to support localization
While trying the setup UI on Backdrop, I found that changing the language
led to a crash.  The problem: it needed to load the `l10n` files, which led
to more path resolution, which led to failures - because `civicrm.settings.php`
hasn't been loaded. (It doesn't exist yet!)

Fortunately, most of the needful information is available.  The
`Setup::init()` should have found the info and put it into `$model`.  We
just need to load the configuration into the right globals/constants.