setup/res/template.css - Twiddle column width
The aim is to have 3 columns (each 1/3 of the screen width) and allow for
some padding in between.  Unfortunately, the current formulation is futzy.
In my browser, when viewing on Backdrop, the columns are just slightly too
wide (causing them to wrap prematurely).

There must be a better way to do this, but it's not my main issue right now.

Twiddling is better than status quo.