dev/translation#48 Provide fall back function for if php-intl is not present
This provides a softer migration to the use of Brick Money, permitting a fallback to the old method if php-intl is

not present. We have a system check to encourage people to install it as step one of trying to ensure people have it.

Deprecation warnings, upgrade notices can be added later