Status Checks - Use more specific label regarding "Domain"/"Organization" check
Before
------

Suppose you install CiviCRM anew.  The installer shows a bulleted list of
things to do.  One of those takes you to another, much longer list of things
to do.  And *then* there's a popup on the sidebar which says you need to
"Complete Setup".

You say to yourself, "I thought that's what I was doing!" But you click it anyway.

That takes you to a page with another list of things to do, one of which is
to "Complete Setup".  The text and link clarify that "Complete Setup"
actually means *set the name and address of the organization*.

After
-----

There are still several different pages telling you what to do after setup.

But at least the label is more precise. :)